fix: исправления по итогам тестирования

+++ src/main/java/ru/yandex/practicum/filmorate/service/review/impl/ReviewServiceImpl.java
@@ -5,8 +5,10 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.dao.review.ReviewStorage;
+import ru.yandex.practicum.filmorate.exception.AlreadyExistsException;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.exception.ReviewNotFoundException;
+import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Review;
 import ru.yandex.practicum.filmorate.service.film.FilmService;
 import ru.yandex.practicum.filmorate.service.review.ReviewService;
@@ -52,7 +54,6 @@ public class ReviewServiceImpl implements ReviewService {
 
     @Override
     public Review updateReview(Review review) {
-        isReviewExist(review.getReviewId());
         userService.isExistsIdUser(review.getUserId());
         filmService.isExistsIdFilm(Long.valueOf(review.getFilmId()).intValue());
         Review updatedReview = reviewStorage.updateReview(review);
@@ -63,11 +64,14 @@ public class ReviewServiceImpl implements ReviewService {
     @Override
     public String addLike(Long id, Long userId) {
         userService.isExistsIdUser(userId);
-        isLikeExist(id, userId);
-        reviewStorage.addLike(id, userId);
+        if (reviewStorage.isLikeExist(id, userId)) {
+            log.warn("Like from user with id={} for the review with id={} already exists", userId, id);
+            throw new AlreadyExistsException(String.format("Like from user with id=%d for the review with id=%d " +
+                            "already exists",
+                    userId, id));
+        }
 
-        Review review = reviewStorage.getReviewById(id);
-        review.setUseful(review.getUseful() + 1);
+        reviewStorage.addLike(id, userId);
 
         log.info("Like from user id={} for the review id={} was added", userId, id);
         return String.format("Like from user id=%d for the review id=%d was added", userId, id);
@@ -76,12 +80,14 @@ public class ReviewServiceImpl implements ReviewService {
     @Override
     public String addDislike(Long id, Long userId) {
         userService.isExistsIdUser(userId);
-        isDislikeExist(id, userId);
-        reviewStorage.addDislike(id, userId);
+        if (reviewStorage.isDislikeExist(id, userId)) {
+            log.warn("Dislike from user with id={} for the review with id={} already exists", userId, id);
+            throw new AlreadyExistsException(String.format("Dislike from user with id=%d for the review with id=%d " +
+                            "already exists",
+                    userId, id));
+        }
 
-        Review review = reviewStorage.getReviewById(id);
-        review.setUseful(review.getUseful() - 1);
-        reviewStorage.updateReview(review);
+        reviewStorage.addDislike(id, userId);
 
         log.info("Dislike from user id={} for the review id={} was added", userId, id);
         return String.format("Dislike from user id=%d for the review id=%d was added", userId, id);
@@ -90,12 +96,13 @@ public class ReviewServiceImpl implements ReviewService {
     @Override
     public String deleteLike(Long id, Long userId) {
         userService.isExistsIdUser(userId);
-        isLikeExist(id, userId);
-        reviewStorage.deleteLike(id, userId);
+        if (!reviewStorage.isLikeExist(id, userId)) {
+            log.warn("Like from user with id={} for the review with id={} not found", userId, id);
+            throw new NotFoundException(String.format("Like from user with id=%d for the review with id=%d not found",
+                    userId, id));
+        }
 
-        Review review = reviewStorage.getReviewById(id);
-        review.setUseful(review.getUseful() - 1);
-        reviewStorage.updateReview(review);
+        reviewStorage.deleteLike(id, userId);
 
         log.info("Like from user id={} for the review id={} was deleted", userId, id);
         return String.format("Like from user id=%d for the review id=%d was deleted", userId, id);
@@ -104,12 +111,13 @@ public class ReviewServiceImpl implements ReviewService {
     @Override
     public String deleteDislike(Long id, Long userId) {
         userService.isExistsIdUser(userId);
-        isDislikeExist(id, userId);
-        reviewStorage.deleteDislike(id, userId);
+        if (!reviewStorage.isDislikeExist(id, userId)) {
+            log.warn("Dislike from user with id={} for the review with id={} not found", userId, id);
+            throw new NotFoundException(String.format("Dislike from user with id=%d for review with id=%d not found",
+                    userId, id));
+        }
 
-        Review review = reviewStorage.getReviewById(id);
-        review.setUseful(review.getUseful() + 1);
-        reviewStorage.updateReview(review);
+        reviewStorage.deleteDislike(id, userId);
 
         log.info("Dislike from user id={} for the review id={} was deleted", userId, id);
         return String.format("Dislike from user id=%d for the review id=%d was deleted", userId, id);
@@ -123,29 +131,29 @@ public class ReviewServiceImpl implements ReviewService {
         return String.format("Review with id=%d deleted", id);
     }
 
-    private void isReviewExist(Long reviewId) {
+    @Override
+    public void isReviewValid(Review review) {
+        if (review.getUserId() == 0) {
+            log.warn("Incorrect userId={} was passed when creating the review: ", review.getUserId());
+            throw new ValidationException("review must have userId");
+        }
+
+        if (review.getFilmId() == 0) {
+            log.warn("Incorrect filmId={} was passed when creating the review: ", review.getFilmId());
+            throw new ValidationException("review must have filmId");
+        }
+
+        if (review.getIsPositive() == null) {
+            log.warn("Incorrect isPositive={} was passed when creating the review: ", review.getIsPositive());
+            throw new ValidationException("review must have isPositive parameter");
+        }
+    }
+
+    public void isReviewExist(Long reviewId) {
         boolean isExist = reviewStorage.isReviewExist(reviewId);
         if (!isExist) {
             log.warn("Review with id={} not found", reviewId);
             throw new ReviewNotFoundException(String.format("Review with id=%d not found", reviewId));
-        }
-    }
-
-    private void isLikeExist(Long reviewId, Long userId) {
-        boolean isExist = reviewStorage.isLikeExist(reviewId, userId);
-        if (!isExist) {
-            log.warn("Like from user with id={} for the review with id={} not found", userId, reviewId);
-            throw new NotFoundException(String.format("Like from user with id=%d for the review with id=%d not found",
-                    userId, reviewId));
-        }
-    }
-
-    private void isDislikeExist(Long reviewId, Long userId) {
-        boolean isExist = reviewStorage.isDislikeExist(reviewId, userId);
-        if (!isExist) {
-            log.warn("Dislike from user with id={} for the review with id={} not found", userId, reviewId);
-            throw new NotFoundException(String.format("Dislike from user with id=%d for review with id=%d not found",
-                    userId, reviewId));
         }
     }
 }
